+++ src/app/[lng]/(sub)/login/page.tsx
@@ -1,0 +1,5 @@
+import Login from './Login';
+
+export default async function page() {
+  return <Login />;
+}
