--- src/app/[lng]/(sub)/token/page.tsx
@@ -1,5 +1,0 @@
-import Token from './Token';
-
-export default function page() {
-  return <Token />;
-}
